+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.2.6] - 2023-05-24
 
 - Add support for sharding via the parameter ``shard_level`` to ``Bigdict.new``.
+- Change default ``map_size`` to 64 MB.
 
 
 ## [0.2.5] - 2023-05-22